Remove conflicting interop structs (#7561)
Removed our `RECT`, `HDC`, etc structs in favor of cswin32. This necessitated changing all of the existing P/Invokes back to `[DllImport]` as the `[LibaryImport]` doesn't work with the source generated types.

- Remove all `using foo =` workarounds.
- Push forward to cswin32 daily build to address bool cast issue
- New cswin32 uses System.Drawing types where plausible (`POINT` -> `Point`, as we currently do).

# Conflicts:
#	eng/Versions.props
#	src/System.Windows.Forms.Design/src/System/Windows/Forms/Design/DesignerUtils.cs
#	src/System.Windows.Forms.Primitives/src/Interop/Gdi32/Interop.CreatePenScope.cs
#	src/System.Windows.Forms/src/System/Windows/Forms/ControlPaint.cs
#	src/System.Windows.Forms/src/System/Windows/Forms/NativeWindow.cs

+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewTextBoxEditingControl.cs
@@ -262,7 +262,7 @@ namespace System.Windows.Forms
 
         protected override bool ProcessKeyEventArgs(ref Message m)
         {
-            switch ((Keys)m.WParamInternal)
+            switch ((Keys)(nint)m.WParamInternal)
             {
                 case Keys.Enter:
                     if (m.MsgInternal == User32.WM.CHAR
