+++ src/System.Windows.Forms/src/System/Windows/Forms/MaskedTextBox.cs
@@ -2888,7 +2888,7 @@ namespace System.Windows.Forms
         private void WmPrint(ref Message m)
         {
             base.WndProc(ref m);
-            if (((User32.PRF)m.LParamInternal & User32.PRF.NONCLIENT) != 0
+            if (((User32.PRF)(nint)m.LParamInternal & User32.PRF.NONCLIENT) != 0
                 && Application.RenderWithVisualStyles && BorderStyle == BorderStyle.Fixed3D)
             {
                 using Graphics g = Graphics.FromHdc(m.WParamInternal);
