+++ src/System.Windows.Forms/src/System/Windows/Forms/SplitContainer.SplitContainerMessageFilter.cs
@@ -24,7 +24,7 @@ namespace System.Windows.Forms
                     return false;
                 }
 
-                if ((m.MsgInternal == User32.WM.KEYDOWN && (Keys)m.WParamInternal == Keys.Escape)
+                if ((m.MsgInternal == User32.WM.KEYDOWN && (Keys)(nint)m.WParamInternal == Keys.Escape)
                     || (m.MsgInternal == User32.WM.SYSKEYDOWN))
                 {
                     // Notify that splitMOVE was reverted. This is used in ONKEYUP.
